* libtiff/tif_config.vc.h: Make adjustments to match the new definitions that configure produces, including for WIN64.  Still needs to be tested.
'lld' is not assured by the run-time DLLs and so GCC warns.
Add TIFF_SIZE_T and TIFF_SIZE_FORMAT to provide a type definition
and printf format specifier to deal with printing values of
'size_t' type.  In particular, this was necessary for WIN64.
Added a configure test for if the system headers provide 'optarg'
(normal case) and block out the many explicit 'extern' statements
in the utilities.  This was found to be necessary under Windows
when getopt is in a DLL and the symbols are already imported with
dllimport via standard header files.

+++ contrib/iptcutil/iptcutil.c
@@ -1,4 +1,4 @@
-/* $Id: iptcutil.c,v 1.10 2015-05-29 13:43:47 bfriesen Exp $ */
+/* $Id: iptcutil.c,v 1.11 2015-06-21 01:09:09 bfriesen Exp $ */
 
 #include "tif_config.h"
 
@@ -801,11 +801,11 @@ void chstore(char *string,int max,char ch)
         switch(_p_flag&3)
           {
           case 1: 	    /* convert to upper */
-            c=toupper(ch);
+            c=toupper((int) ch);
             break;
 
           case 2: 	    /* convert to lower */
-            c=tolower(ch);
+            c=tolower((int) ch);
             break;
 
           default:	    /* use as is */
